added files upload/download to mike-db api

+++ src/main/java/org/maggus/mikedb/services/PersistenceService.java
@@ -1,8 +1,11 @@
 package org.maggus.mikedb.services;
 
 import lombok.extern.java.Log;
+import org.maggus.mikedb.data.FileItem;
+import org.maggus.mikedb.data.FileItemStream;
 
 import java.io.File;
+import java.io.FileOutputStream;
 import java.io.FilenameFilter;
 import java.io.IOException;
 import java.nio.file.Path;
@@ -12,9 +15,12 @@ import java.util.Map;
 
 @Log
 public class PersistenceService {
+    public static final String DB_VALUE_EXT = ".db";
+    public static final String DB_FILE_EXT = ".dbfile";
 
     private final DbService db;
     private Map<String, File> valuesFiles = new HashMap<>();
+    private Map<String, File> binaryFiles = new HashMap<>();
 
     protected PersistenceService(DbService db) {
         this.db = db;
@@ -35,7 +41,7 @@ public class PersistenceService {
         return dbDir.listFiles(new FilenameFilter() {
             @Override
             public boolean accept(File dir, String name) {
-                return name.endsWith(".db");
+                return name.endsWith(DB_VALUE_EXT) || name.endsWith(DB_FILE_EXT);
             }
         });
     }
@@ -45,9 +51,27 @@ public class PersistenceService {
         if (file != null) {
             return file;
         }
-        file = new File(getDbDir(), name + ".db");
+        file = new File(getDbDir(), name + DB_VALUE_EXT);
         valuesFiles.put(name, file);
         return file;
+    }
+
+    private File getDbBinaryFile(String name, String type) throws IOException {
+        File file = binaryFiles.get(name);
+        if (file != null) {
+            return file;
+        }
+        String fName = name;
+        if (type != null && !fName.endsWith("." + type)) {
+            fName += "." + type;
+        }
+        file = new File(getDbDir(), fName + DB_FILE_EXT);
+        binaryFiles.put(name, file);
+        return file;
+    }
+
+    private File findDbBinaryFile(String name) throws IOException {
+        return binaryFiles.get(name);
     }
 
     /**
@@ -62,11 +86,16 @@ public class PersistenceService {
         if (files != null) {
             for (File file : files) {
                 String fName = file.getName();
-                if (!file.isFile() || !fName.endsWith(".db")) {
-                    continue;
+                if (file.isFile() && fName.endsWith(DB_VALUE_EXT)) {
+                    String key = fName.substring(0, fName.length() - DB_VALUE_EXT.length());
+                    items.put(key, JsonUtils.fileToObject(file));
+                    valuesFiles.put(key, file);
+                } else if (file.isFile() && fName.endsWith(DB_FILE_EXT)) {
+                    FileItem fileItem = new FileItem(file);
+                    String key = fileItem.getName();    // name is a key
+                    items.put(key, fileItem);
+                    binaryFiles.put(key, file);
                 }
-                String key = fName.substring(0, fName.length() - 3);
-                items.put(key, JsonUtils.fileToObject(file));
             }
         }
     }
@@ -82,8 +111,34 @@ public class PersistenceService {
         if (value == null) {
             //delete old value files if any
             File file = getDbValueFile(key);
-            if (file.isFile()) {
+            if (file != null && file.isFile()) {
                 file.delete();
+            }
+            File binaryFile = findDbBinaryFile(key);
+            if (binaryFile != null && binaryFile.isFile()) {
+                binaryFile.delete();
+            }
+        } else if (value instanceof FileItemStream) {
+            FileItemStream fi = (FileItemStream) value;
+            FileItem fileItem = fi.getFileItem();
+            // augment the key with the file extension to get the final db file name
+            File file = getDbBinaryFile(key, fileItem.getType());
+            // store full local file path as file name
+            fileItem.setFileName(file.getAbsolutePath());
+            // finally write to a db file
+            FileOutputStream out = null;
+            try {
+                out = new FileOutputStream(file);
+                int read = 0;
+                byte[] bytes = new byte[1024];
+                while ((read = fi.getFileInputStream().read(bytes)) != -1) {
+                    out.write(bytes, 0, read);
+                }
+            } finally {
+                if (out != null) {
+                    out.flush();
+                    out.close();
+                }
             }
         } else {
             String json = JsonUtils.objectToString(value);
@@ -120,6 +175,9 @@ public class PersistenceService {
 
 
     public static boolean isValidName(String name) {
+        if (name.contains("\\") || name.contains("/")) {
+            return false;
+        }
         File f = new File(name);
         try {
             f.getCanonicalPath();
